+++ changelog.md
@@ -1,3 +1,8 @@
+# 2.0.2
+
+- Added `rateLimitMs` option to the `FetchHttpClient`.
+  - This allows you to restrict how often the client is allowed to make calls, which can be useful when consuming an API with call rate limitations.
+
 # 2.0.1
 
 - Add repo information.